refactor(structure): modularize 3d features and align e2e targets
- reorganize app/features/3d into hello-world, hello-shader-world, random-walk-world modules

- move tests and fixtures into per-target directories

- add test:e2e:random-walk-world package script

- update routes, traceability paths, and docs to new structure
