+++ app/features/3d/hello-world/camera-actions.ts
