+++ app/features/3d/random-walk-world/random-walk-toroidal-boundary.pseudocode.ts
@@ -3,7 +3,7 @@ export type Issue32RequirementId = "CH-001" | "CH-003";
 export type RequirementLogicObligation = {
   id: Issue32RequirementId;
   obligation: string;
-  owningPseudocodeLocus: "app/features/3d/random-walk-toroidal-boundary.pseudocode.ts";
+  owningPseudocodeLocus: "app/features/3d/random-walk-world/random-walk-toroidal-boundary.pseudocode.ts";
   requirementCases: readonly string[];
 };
 
@@ -12,7 +12,7 @@ export const ISSUE_32_LOGIC_OBLIGATIONS: readonly RequirementLogicObligation[] =
     id: "CH-001",
     obligation:
       "When the third left menu option is selected, initialize and expose a random-walk dot-cloud scene with label and parameter controls.",
-    owningPseudocodeLocus: "app/features/3d/random-walk-toroidal-boundary.pseudocode.ts",
+    owningPseudocodeLocus: "app/features/3d/random-walk-world/random-walk-toroidal-boundary.pseudocode.ts",
     requirementCases: [
       "menu.third-option.selection.initializes-random-walk-dot-cloud-scene",
       "menu.third-option.label.identifies-random-walk-sphere-visualization",
@@ -23,7 +23,7 @@ export const ISSUE_32_LOGIC_OBLIGATIONS: readonly RequirementLogicObligation[] =
     id: "CH-003",
     obligation:
       "When a dot exits a cube boundary on any axis, wrap position to the opposite boundary while preserving velocity continuity.",
-    owningPseudocodeLocus: "app/features/3d/random-walk-toroidal-boundary.pseudocode.ts",
+    owningPseudocodeLocus: "app/features/3d/random-walk-world/random-walk-toroidal-boundary.pseudocode.ts",
     requirementCases: [
       "toroidal-boundary.positive-axis.crossing.wraps-to-opposite-side",
       "toroidal-boundary.negative-axis.crossing.wraps-to-opposite-side",
@@ -72,14 +72,14 @@ export type DeriveToroidalWrapTransition = (
 
 export type Issue32PseudocodeLocus = {
   requirementId: Issue32RequirementId;
-  owner: "app/features/3d/random-walk-toroidal-boundary.pseudocode.ts";
+  owner: "app/features/3d/random-walk-world/random-walk-toroidal-boundary.pseudocode.ts";
   pseudocodeArtifacts: readonly string[];
 };
 
 export const ISSUE_32_PSEUDOCODE_LOCI: readonly Issue32PseudocodeLocus[] = [
   {
     requirementId: "CH-001",
-    owner: "app/features/3d/random-walk-toroidal-boundary.pseudocode.ts",
+    owner: "app/features/3d/random-walk-world/random-walk-toroidal-boundary.pseudocode.ts",
     pseudocodeArtifacts: [
       "DeriveRandomWalkMenuSelectionFlow",
       "MenuSelectionFlowInput",
@@ -88,7 +88,7 @@ export const ISSUE_32_PSEUDOCODE_LOCI: readonly Issue32PseudocodeLocus[] = [
   },
   {
     requirementId: "CH-003",
-    owner: "app/features/3d/random-walk-toroidal-boundary.pseudocode.ts",
+    owner: "app/features/3d/random-walk-world/random-walk-toroidal-boundary.pseudocode.ts",
     pseudocodeArtifacts: [
       "DeriveToroidalWrapTransition",
       "DotKinematics",
