+++ app/routes/hello-shader-world.tsx
@@ -1,4 +1,4 @@
-import { HelloShaderWorldPage } from "~/features/3d/hello-shader-world-page";
+import { HelloShaderWorldPage } from "~/features/3d/hello-shader-world/hello-shader-world-page";
 
 export default function HelloShaderWorldRoute() {
   return <HelloShaderWorldPage />;
