+++ app/routes/random-walk-world.tsx
@@ -1,4 +1,4 @@
-import { RandomWalkWorldPage } from "~/features/3d/random-walk-world-page";
+import { RandomWalkWorldPage } from "~/features/3d/random-walk-world/random-walk-world-page";
 
 /** Issue #32 route placement mapping: CH-001. */
 export default function RandomWalkWorldRoute() {
