+++ tests/hello-shader-world/contracts/hello-shader-world.frame-000.txt
