+++ tests/hello-shader-world/contracts/hello-shader-world.frame-030.txt
