+++ tests/hello-shader-world/contracts/hello-shader-world.frame-060.txt
