+++ tests/hello-shader-world/contracts/hello-shader-world.frame-090.txt
