+++ tests/hello-world/contracts/hello-world.camera.step-01.txt
