+++ tests/hello-world/contracts/hello-world.camera.step-02.txt
