+++ tests/hello-world/contracts/hello-world.camera.step-03.txt
