+++ tests/hello-world/contracts/hello-world.camera.step-04.txt
