+++ tests/hello-world/contracts/hello-world.camera.step-05.txt
