+++ tests/hello-world/contracts/hello-world.camera.step-06.txt
