+++ tests/hello-world/contracts/hello-world.camera.step-07.txt
