+++ tests/hello-world/contracts/hello-world.camera.step-08.txt
