+++ tests/hello-world/contracts/hello-world.camera.step-09.txt
