+++ tests/hello-world/contracts/hello-world.camera.step-10.txt
