+++ tests/hello-world/contracts/hello-world.initial.txt
