+++ tests/random-walk-world/contracts/random-walk-world.ms-000.txt
