+++ tests/random-walk-world/contracts/random-walk-world.ms-072.txt
