+++ tests/random-walk-world/contracts/random-walk-world.ms-144.txt
