+++ tests/random-walk-world/contracts/random-walk-world.ms-216.txt
