+++ tests/random-walk-world/contracts/random-walk-world.ms-288.txt
