+++ tests/random-walk-world/contracts/random-walk-world.ms-360.txt
