+++ tests/random-walk-world/random-walk-world.contract.spec.ts
